除算メソッドをfloatに対応さす

+++ src/junit/tutorial/Calculator.java
@@ -6,6 +6,6 @@ public class Calculator {
     }
 
     public int divide(int x, int y) {
-        return x / y;
+        return (float)x / (float)y;
     }
 }
